Ajout commandes diverses + tentative d'ajout de fonction externe
