+++ src/index.js
@@ -1,9 +1,8 @@
 const client = require("./client/client.js");
-const message = require("./utils/message.js");
-const command = require("./utils/command.js");
+const command = require("./services/command.js");
 
 client.on ("ready", () => {
     console.log("🌤 Weather Bot Online ☁");
 });
 
-client.on("message", message);
+client.on("command", command);
