+++ src/utils/cities.json
@@ -1,0 +1,11 @@
+{
+    "Nantes":"http://di262mgurvkjm.cloudfront.net/777f6424-f59f-4e7b-8705-b9acbc8c6c2d/bouffay-Nantes-0304_uxga.jpg",
+    "Paris":"https://evasion-online.com/images/2016/02/tourisme-paris.jpg",
+    "Brest":"https://www.breizh.watch/wp-content/uploads/2012/05/Chateau_de_Brest_et_Tour_Tanguy.jpgRe",
+    "Rennes":"https://www.letelegramme.fr/ar/imgproxy.php/images/2016/11/28/la-municipalite-veut-redonner-toute-sa-place-a-la-vilaine-a_3190044.jpg?article=20161128-1011308702&aaaammjj=20161128",
+    "Morlaix":"https://www.familiscope.fr/assets/fiches/47000/47305-pont-morlaix.jpg",
+    "Grenoble":"https://www.orientaction-groupe.com/wp-content/uploads/2019/05/GRENOBLE.jpg",
+    "Lille":"http://img.20mn.fr/fgGzwhcGSFKdl8dtBawfAg/2048x1536-fit_beffroi-mairie-lille.jpg",
+    "Montpellier":"https://uploads.lebonbon.fr/source/2019/october/gbu8bl33j7_2_1293.jpg"
+
+}
